Only alter instance count if we need to
There are some Resources that set `instances_best_fit` to value larger
than one, but `instances` to zero. In these cases, we will incorrectly
set `instances` to 1. This causes user confusion and may cause an
incorrect configuration.

This new logic should only set `instances` to 1 if the PAS tile
configuration would have deployed more than one instance of the Resource
in question.

It also handles the case where `instances` is set to the empty string.
This corresponds to the "Automatic" setting in the Resource config,
whose value comes from the `instances_best_fit` key.

[#163839863]