Revert "Grab stemcells from bosh.io"
The Bosh-Core team accidentally published the stemcells we required for
older PCFs to bosh.io. Unfortunately, the only way we are going to be
able to get the trusty stemcells we require is through pivnet.

The Bosh-Core team is working to give us temporary access to the
stemcells required to install 2.x PCFs. They should have this done long
before EOD today.

[#166048896]