Check that desired stemcell file exists
TAS environment creation broke after the latest version of a GCP
stemcell was removed from the product files for that stemcell's PivNet
release entry. The previous implementation assumed that the presence of
a stemcell release guaranteed the presence of a stemcell file for the
desired IaaS.

This change will check that the stemcell version specified in the tile's
manifest and the latest stemcell version in the same major line (if one
exists) actually have a stemcell file for the desired IaaS; it prefers
the newer version. It doesn't handle the case where there is a usable
version between the latest and the tile's version.

An improvemd implemenation would be to loop over the versions from the latest
version down to the version in the tile, terminating the loop once it finds a
stemcell version with a file for the desired IaaS.