Add DNS for `login.sys` and `uaa.sys`
Recently we keep seeing the server does not exist for `login.sys.ENV.cf-app.com` and `uaa.sys.ENV.cf-app.com`.
It looks like it is a GCP BUG. Because we've already had `*.sys.ENV.cf-app.com`, it should cover the host `login` and `uaa`.
From my observation, anything except `login` and `uaa` could be resolved correctly. For now we have to explictly
state the DNS for `login` and `uaa`.

[#177709735]

Signed-off-by: Bin Ju <bju@pivotal.io>